Extract reporting batch contracts
