+++ src/app/routers/reporting_batch_status.py
@@ -2,7 +2,7 @@ from typing import Annotated
 
 from fastapi import APIRouter, Path
 
-from app.contracts.reporting import BATCH_STATUS_RESPONSE_EXAMPLE, BatchStatusResponse
+from app.contracts.reporting_batches import BATCH_STATUS_RESPONSE_EXAMPLE, BatchStatusResponse
 from app.middleware.correlation import correlation_id_var
 from app.routers.reporting_context import ReportingCallerContext
 from app.routers.reporting_errors import report_batch_error_response
